refactor(storage): move update_mcp_server_tools to mcp_servers module
Address review feedback: method was misplaced in skills.rs during split.
