+++ crates/server/src/storage/memory/skills.rs
@@ -5,7 +5,7 @@ use super::InMemoryDatabase;
 use super::matches_search_tokens;
 use anyhow::Result;
 use anyhow::anyhow;
-use everruns_core::{ImageId, McpServerId, SkillId};
+use everruns_core::{ImageId, SkillId};
 use uuid::Uuid;
 
 impl InMemoryDatabase {
@@ -100,26 +100,6 @@ impl InMemoryDatabase {
             .take(limit as usize)
             .collect();
         Ok(result)
-    }
-
-    pub async fn update_mcp_server_tools(
-        &self,
-        org_id: i64,
-        id: Uuid,
-        input: UpdateMcpServerTools,
-    ) -> Result<Option<McpServerRow>> {
-        let id = McpServerId::from_uuid(id);
-        let mut servers = self.mcp_servers.write();
-        if let Some(server) = servers.get_mut(&id) {
-            if server.org_id != org_id {
-                return Ok(None);
-            }
-            server.cached_tools = input.cached_tools;
-            server.tools_cached_at = Some(Self::now());
-            server.updated_at = Self::now();
-            return Ok(Some(server.clone()));
-        }
-        Ok(None)
     }
 
     // ============================================
